remove unused function

+++ DateTime.cpp
@@ -6,7 +6,7 @@
 #define SECONDS_FROM_1970_TO_2000 946684800
 
 //has to be const or compiler compaints
-const uint16_t daysInMonth [] PROGMEM = {31,28,31,30,31,30,31,31,30,31,30,31};
+const uint8_t daysInMonth [] PROGMEM = {31,28,31,30,31,30,31,31,30,31,30,31};
 
 // number of days since 2000/01/01, valid for 2001..2178
 static uint16_t date2days(uint16_t y, uint16_t m, uint16_t d) {
